@@ -16,7 +16,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router, ParamMap } from '@angular/router';
 import { trigger, state, animate, transition, style } from '@angular/animations';
-import 'rxjs/add/operator/switchMap';
+import { Observable } from 'rxjs';
 import { QuoteIb } from '../common/quoteIb';
 import { ItbitService} from '../services/itbit.service';
 import { CommonUtils } from '../common/commonUtils';
@@ -47,13 +47,10 @@ export class IbdetailComponent implements OnInit {
     constructor(private route: ActivatedRoute, private router: Router, private serviceIb: ItbitService) { }
 
     ngOnInit() {
-        this.route.paramMap
-            .switchMap((params: ParamMap) => {
-                this.currPair = this.utils.getCurrpairName(params.get('currpair'));
-                return this.serviceIb.getCurrentQuote(params.get('currpair'));})
+        this.currPair = this.route.snapshot.paramMap.get('currpair');
+        this.serviceIb.getCurrentQuote(this.currPair)
             .subscribe(quote => this.currQuote = quote);
-        this.route.paramMap
-            .switchMap((params: ParamMap) => this.serviceIb.getTodayQuotes(params.get('currpair')))
+        this.serviceIb.getTodayQuotes(this.currPair)
             .subscribe(quotes => {
                 this.todayQuotes = quotes;
                 this.chartlabels = this.todayQuotes.map(quote => new Date(quote.createdAt).getHours().toString());
@@ -64,14 +61,13 @@ export class IbdetailComponent implements OnInit {
     changeTf() {
         this.chartdata = [];
         this.chartlabels = [];
-        this.route.paramMap
-        .switchMap((params: ParamMap) => {           
-            if(this.timeframe === this.utils.timeframes[1]) return this.serviceIb.get7DayQuotes(params.get('currpair'));
-            if(this.timeframe === this.utils.timeframes[2]) return this.serviceIb.get30DayQuotes(params.get('currpair'));
-            if(this.timeframe === this.utils.timeframes[3]) return this.serviceIb.get90DayQuotes(params.get('currpair')) 
-                else return this.serviceIb.getTodayQuotes(params.get('currpair'));
-        })            
-        .subscribe(quotes => { 
+        const currpair = this.route.snapshot.paramMap.get('currpair');
+        let quoteObserv: Observable<QuoteIb[]>;
+            if(this.timeframe === this.utils.timeframes[1]) quoteObserv = this.serviceIb.get7DayQuotes(currpair);
+            if(this.timeframe === this.utils.timeframes[2]) quoteObserv = this.serviceIb.get30DayQuotes(currpair);
+            if(this.timeframe === this.utils.timeframes[3]) quoteObserv = this.serviceIb.get90DayQuotes(currpair) 
+                else quoteObserv = this.serviceIb.getTodayQuotes(currpair);
+        quoteObserv.subscribe(quotes => { 
             this.todayQuotes = quotes;
             if(this.timeframe === this.utils.timeframes[2] || this.timeframe === this.utils.timeframes[3]) 
                 this.chartlabels = this.todayQuotes.map(quote => new Date(quote.createdAt).getUTCDate().toString())            
