@@ -22,10 +22,8 @@ import { QuoteBs } from '../common/quoteBs';
 import { QuoteCb } from '../common/quoteCb';
 import { QuoteIb } from '../common/quoteIb';
 import { QuoteBf } from '../common/quoteBf';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of, Subject } from 'rxjs';
 import {DataSource, CollectionViewer} from '@angular/cdk/collections';
-import 'rxjs/add/observable/of';
-import {Subject} from 'rxjs/Subject';
 import { Router } from '@angular/router';
 import { CommonUtils } from "../common/commonUtils";
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
