@@ -16,11 +16,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { PlatformLocation } from '@angular/common';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/observable/throw';
+import { Observable } from 'rxjs';
+import { catchError, map, tap } from 'rxjs/operators';
 import { QuoteBs } from '../common/quoteBs';
 import { QuoteCb } from '../common/quoteCb';
 import { QuoteIb } from '../common/quoteIb';
@@ -39,26 +36,26 @@ export class ItbitService {
     }
 
     getCurrentQuote(currencypair: string): Observable<QuoteIb> {
-        return this.http.get(this._itbit+'/'+currencypair+'/current', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<QuoteIb>(this._itbit+'/'+currencypair+'/current', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<QuoteIb>('getCurrentQuote')));
     }
     
     getTodayQuotes(currencypair: string): Observable<QuoteIb[]> {
-        return this.http.get(this._itbit+'/'+currencypair+'/today', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<QuoteIb[]>(this._itbit+'/'+currencypair+'/today', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<QuoteIb[]>('getTodayQuotes')));
     }
     
     get7DayQuotes(currencypair: string): Observable<QuoteIb[]> {
-        return this.http.get(this._itbit+'/'+currencypair+'/7days', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<QuoteIb[]>(this._itbit+'/'+currencypair+'/7days', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<QuoteIb[]>('get7DayQuotes')));
     }
     
     get30DayQuotes(currencypair: string): Observable<QuoteIb[]> {
-        return this.http.get(this._itbit+'/'+currencypair+'/30days', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<QuoteIb[]>(this._itbit+'/'+currencypair+'/30days', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<QuoteIb[]>('get30DayQuotes')));
     }
     
     get90DayQuotes(currencypair: string): Observable<QuoteIb[]> {
-        return this.http.get(this._itbit+'/'+currencypair+'/90days', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<QuoteIb[]>(this._itbit+'/'+currencypair+'/90days', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<QuoteIb[]>('get30DayQuotes')));
     }
     
     getOrderbook(currencypair: string): Observable<OrderbookIb> {
-        return this.http.get(this._itbit+'/'+currencypair+'/orderbook/', this._reqOptionsArgs).catch(this._utils.handleError);
+        return this.http.get<OrderbookIb>(this._itbit+'/'+currencypair+'/orderbook/', this._reqOptionsArgs).pipe(catchError(this._utils.handleError<OrderbookIb>('getOrderbook')));
     }
 }
