@@ -14,15 +14,27 @@
    limitations under the License.
  */
 import { Http, Response, RequestOptionsArgs, Headers } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of } from 'rxjs';
 
 export class Utils {
     
-    handleError( error: Response ) {
-        // in a real world app, we may send the error to some remote logging infrastructure
-        // instead of just logging it to the console
-        console.error( JSON.stringify( error ) );
-        return Observable.throw( error );
-    }
+    public handleError<T> (operation = 'operation', result?: T) {
+        return (error: any): Observable<T> => {
+     
+          // TODO: send the error to remote logging infrastructure
+          console.error(error); // log to console instead
+     
+          // TODO: better job of transforming error for user consumption
+          this.log(`${operation} failed: ${error.message}`);
+     
+          // Let the app keep running by returning an empty result.
+          return of(result as T);
+        };
+      }
+     
+      /** Log a HeroService message with the MessageService */
+      private log(message: string) {
+       console.log(message);
+      }
   
 }
